requirements_test.txt
pep8==1.7.0
flake8==2.5.4
mock==1.3.0
coverage==4.0.3
coveralls==1.1
git+https://github.com/openstack/bandit.git   # PyPi version is out of date
pytest==3.0.5
pytest_cov==2.4.0
pytest_django==3.1.2
